input: refactor input system
Add missing comments and simplify the logic.

Use the same composite PointInput for touch, point, and mouse.
That allows the same control function to be used for all inputs.

Fix composite errors in the InputSystem. It seems that it doesn't
work correctly when property is written in camelCase. Rename to Id.

Introduce DEBUG_LOG_INPUT that disables input related log output.