lvlmgr: fix block goals label not updating
The bug exists only when the game runs outside of the Editor.
The label is not re-rendered after update in LateUpdate.
Fix is to re-set the label on each block goal check.

Also correct the block creation when there are no more
blocks on the central. Should happen after a match.